Fixed #647 window open/close flickering
Opening and closing a set of windows with at least one other window
active caused flickering if a window as closed. Reason was a missing
check inside `nk_build` that did not skip newly closed windows.

I am not 100% sure but this could also affect #637 but I still needs to
be tested.